requirements.txt
streamlit==1.39.0
requests==2.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.1.2
pytz==2024.2
plotly==5.17.0
